pyproject.toml
[project]
name = "poshdash-agents"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-openai>=0.3",
    "python-dotenv>=1.1",
    "requests>=2.31",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[project.scripts]
agents = "agents.main:main"
